Final Report/Team10/NestWise-main/backend-userAuth/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-jose
argon2-cffi
pydantic[email]
python-dotenv
